refactore tables field for sunc vc in indexer[5042]
